requirements.txt
requests
structlog
urlnorm
sqlalchemy
faststat
ckanapi
pandas
numpy
enum
rarfile
jinja2
tornado
pyyaml
psycopg2
bokeh
pybloom
python-dateutil
scrapy
py-lru-cache
queuelib
rdflib
rdflib-jsonld
tldextract
Twisted
geomet
reppy
dictdiffer
networkx
flask
flask-Caching
flask-restplus
flexx
isoweek
statsmodels
matplotlib